backend/requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.3b0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Restless==0.17.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mimerender==0.6.0
psycopg2==2.8.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-mimeparse==1.6.0
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.10
toml==0.10.0
urllib3==1.25.6
Werkzeug==0.16.0
